refactor(web): extract ONNX metadata parser into onnx_meta
